Added an append node
This node goes after unformatted prompt and has 3 options. 

keyframe interval - Adds keyframes just like the unformatted prompt node. Ignored if append to all is set to true.
append to all - allows for 1 prompt to be appended to all prompts. 
use neg - adds and removes the --neg flag so it can be used for negative prompts and not.